pyproject.toml
[tool.poetry]
name = "desk"
version = "0.1.0"
description = "Custom package for a smart desk."
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"
python-dotenv = "^0.19.2"
paho-mqtt = "^1.6.1"
unsync = "^1.4.0"
more-itertools = "^8.12.0"

[tool.black]
line-length = 120

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
